Update Selenium2Driver.php
The error "invalid argument: 'id' can not be string" is thrown when passing a numeric iframe index to the method switchToIframe, because is casted to string.
 
The package "instaclick/php-webdriver" throws  a 500 error in CurlService. Passing ['id' => integer] here $this->getWebDriverSession()->frame(array('id' => $name)); solves the problem.

That's the reason why i suggest typing switchToIFrame with integer (apart from string and null).